refactor(phoenix): Split up sed code
Let's organise this a bit.
